+++ tests/test_generate_run_request_json_schema.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tests to check if the generated json schema validates run requests.
+"""Tests to check if the generated json schema validates job requests.
 """
 # pylint: disable=redefined-outer-name
 from copy import deepcopy
@@ -20,15 +20,15 @@ import pytest
 from jsonschema import ValidationError, validate
 
 from docs.generate_json_schemas import generate_json_schema
-from iqm_client.iqm_client import RunRequest
+from iqm_client.iqm_client import JobRequest
 
 
 @pytest.fixture
-def sample_valid_run_request(sample_circuit):
+def sample_valid_job_request(sample_circuit):
     """
-    Returns valid example run request.
+    Returns valid example job request.
     """
-    return RunRequest(
+    return JobRequest(
         circuit=sample_circuit,
         settings={},
         shots=1000,
@@ -36,25 +36,25 @@ def sample_valid_run_request(sample_circuit):
     ).dict()
 
 @pytest.fixture
-def sample_invalid_run_request(sample_valid_run_request):
+def sample_invalid_job_request(sample_valid_job_request):
     """
-    Returns invalid example run request.
+    Returns invalid example job request.
     """
-    invalid_run_request = deepcopy(sample_valid_run_request)
-    invalid_run_request['shots'] = 'not_a_number'
-    return invalid_run_request
+    invalid_job_request = deepcopy(sample_valid_job_request)
+    invalid_job_request['shots'] = 'not_a_number'
+    return invalid_job_request
 
-def test_jsonschema_validates_run_requests(sample_valid_run_request):
+def test_jsonschema_validates_job_requests(sample_valid_job_request):
     """
-    Tests that the generated json schema validates valid run requests.
+    Tests that the generated json schema validates valid job requests.
     """
-    json_schema = generate_json_schema(RunRequest, '')
-    validate(schema=json_schema, instance=sample_valid_run_request)
+    json_schema = generate_json_schema(JobRequest, '')
+    validate(schema=json_schema, instance=sample_valid_job_request)
 
-def test_jsonschema_throws_validation_errors(sample_invalid_run_request):
+def test_jsonschema_throws_validation_errors(sample_invalid_job_request):
     """
-    Tests that the generated json schema rejects invalid run requests.
+    Tests that the generated json schema rejects invalid job requests.
     """
-    json_schema = generate_json_schema(RunRequest, '')
+    json_schema = generate_json_schema(JobRequest, '')
     with pytest.raises(ValidationError):
-        validate(schema=json_schema, instance=sample_invalid_run_request)
+        validate(schema=json_schema, instance=sample_invalid_job_request)